chore: refactor bwa aln/mem options into their own modules for compilation speedups
