+++ pyproject.toml
@@ -55,7 +55,7 @@ pytest-cov = ">=2.8.1"
 ruff = "0.4.8"
 setuptools = ">=68.0.0"
 black = ">=24.8.0"  # for readthedocs
-cython = "^3.0.11"
+cython = "^3.0.12"
 coverage = {extras = ["toml"], version = "^7.6.10"}
 
 [tool.poetry.group.docs.dependencies]
